Finished ProductService

+++ src/main/java/com/kyurban/GameBase/product/Product.java
@@ -49,7 +49,7 @@ public class Product {
         this.price = price;
     }
 
-    public long getStock() {
+    public int getStock() {
         return stock;
     }
